Fixed main running before form input is received, and updated lists of skillsets
Added 3 routes in flask file, 1. root route to check if flask runs , 2. Route GET to check if form is submitted, 3. Route POST to reset form submission for running program again.

Updated list of skillsets to be filtered, to cater to more jobs outside of the IT sector.

Removed certain comments as well.